mcap2json: Add mcap bzip2 support
mcap files are often large and therefore stored as compressed files. To
use mcap2json today with bzip2-compressed mcap files, we must first
decompress the files in order to mcap2json to find and process those
files. To avoid the extra storage space requirements and for more
convenience for users, add support for transparent handling of both
bzip2-compressed or non-compressed mcap files.

To do this, we include .mcap.bz2 files in the search for mcap files. We
also create a function to handle computing basename from files that have
.bz2 extensions in addition to their .mcap extension. Additionally, we
create a function that returns which open function to use so we can
transparently support opening either bzip2-compressed or non-compressed
files with the same mcap_reader.